The end of LART sparse implementation attempt
The algorithm as its currently defined in Zhana's paper is impossible to implement using sparse techniques because the distance matrix that is needed to cluster nodes uses small values to dictate whether nodes are in a community or not. This will result in a very dense matrix making everything sparse up to this point futile. If it's possible to use the inverse of this metric (Large distances making clusters) and pruning low values (since now they mean they are far away) then it's possible to continue. However this would change the behaviour of the algorithm and is outside of the scope of my thesis